pyproject.toml
[project]
name = "mcp-futu-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9.22"
dependencies = [
    "flask==2.0.1",
    "futu-api==9.2.5208",
    "gunicorn==20.1.0",
    "python-dotenv==1.0.0",
    "werkzeug==2.0.1",
]

[dependency-groups]
dev = [
    "pre-commit>=4.2.0",
    "ruff>=0.11.11",
]

# Ruff 配置
[tool.ruff]
line-length = 88
target-version = "py39"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by formatter)
    "UP007", # Use `X | Y` for type annotations
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.ruff.lint.isort]
known-first-party = ["mcp_futu_api"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["mcp_futu_api"]

[tool.hatch.build.targets.wheel.force-include]
"." = "."
